HMOS-204:Cash balance calculation to include holdback amount when the wire is past deadline Changing the logic to show min of holdback vs available balance

+++ HedgeMark.Operations.Secure.Middleware/FundAccountManager.cs
@@ -646,7 +646,7 @@ namespace HedgeMark.Operations.Secure.Middleware
                 : ComputeNonPBCashBalances(sendingFundAccountId, valueDate, contextDate);
 
             if (cashbalances != null)
-                cashbalances.HoldBackAmount = fndAccount.HoldbackAmount ?? 0;
+                cashbalances.HoldBackAmount = (decimal)(fndAccount.HoldbackAmount ?? 0);
 
             return cashbalances;
         }
